API: Move PluginMainWidget to its own module
