+++ spyder/plugins/console/widgets/main_widget.py
@@ -26,7 +26,7 @@ from qtpy.QtWidgets import QInputDialog, QLineEdit, QVBoxLayout
 
 # Local imports
 from spyder.api.translations import get_translation
-from spyder.api.widgets import PluginMainWidget
+from spyder.api.widgets.main_widget import PluginMainWidget
 from spyder.api.config.decorators import on_conf_change
 from spyder.app.solver import find_internal_plugins
 from spyder.config.base import DEV, get_debug_level
