+++ spyder/plugins/history/widgets.py
@@ -13,13 +13,12 @@ import sys
 
 # Third party imports
 from qtpy.QtCore import Signal, Slot
-from qtpy.QtGui import QFont
-from qtpy.QtWidgets import QInputDialog, QVBoxLayout, QWidget
+from qtpy.QtWidgets import QVBoxLayout, QWidget
 
 # Local imports
 from spyder.api.config.decorators import on_conf_change
 from spyder.api.translations import get_translation
-from spyder.api.widgets import PluginMainWidget
+from spyder.api.widgets.main_widget import PluginMainWidget
 from spyder.py3compat import is_text_string, to_text_string
 from spyder.utils import encoding
 from spyder.utils.sourcecode import normalize_eols
