+++ spyder/plugins/maininterpreter/container.py
@@ -15,7 +15,7 @@ from qtpy.QtCore import Signal
 
 # Local imports
 from spyder.api.config.decorators import on_conf_change
-from spyder.api.widgets import PluginMainContainer
+from spyder.api.widgets.main_container import PluginMainContainer
 from spyder.plugins.maininterpreter.widgets.status import InterpreterStatus
 from spyder.utils.misc import get_python_executable
