+++ spyder/plugins/pylint/main_widget.py
@@ -24,15 +24,14 @@ import pylint
 from qtpy.compat import getopenfilename
 from qtpy.QtCore import (QByteArray, QProcess, QProcessEnvironment, Qt,
                          Signal, Slot)
-from qtpy.QtWidgets import (QHBoxLayout, QInputDialog, QLabel, QMessageBox,
-                            QSizePolicy, QTreeWidgetItem, QVBoxLayout, QWidget)
+from qtpy.QtWidgets import (QInputDialog, QLabel, QMessageBox, QTreeWidgetItem,
+                            QVBoxLayout)
 
 # Local imports
 from spyder.api.config.decorators import on_conf_change
 from spyder.api.translations import get_translation
-from spyder.api.widgets import PluginMainWidget
+from spyder.api.widgets.main_widget import PluginMainWidget
 from spyder.config.base import get_conf_path, running_in_mac_app
-from spyder.config.gui import is_dark_interface
 from spyder.plugins.pylint.utils import get_pylintrc_path
 from spyder.plugins.variableexplorer.widgets.texteditor import TextEditor
 from spyder.utils.icon_manager import ima
