+++ spyder/plugins/statusbar/container.py
@@ -13,7 +13,7 @@ from qtpy.QtCore import Signal
 
 # Local imports
 from spyder.api.config.decorators import on_conf_change
-from spyder.api.widgets import PluginMainContainer
+from spyder.api.widgets.main_container import PluginMainContainer
 from spyder.plugins.statusbar.widgets.status import (
     ClockStatus, CPUStatus, MemoryStatus
 )
