+++ spyder/plugins/variableexplorer/widgets/main_widget.py
@@ -15,7 +15,7 @@ from qtpy.QtWidgets import QAction, QStackedWidget, QVBoxLayout
 # Local imports
 from spyder.api.config.decorators import on_conf_change
 from spyder.api.translations import get_translation
-from spyder.api.widgets import PluginMainWidget
+from spyder.api.widgets.main_widget import PluginMainWidget
 from spyder.plugins.variableexplorer.widgets.namespacebrowser import (
     NamespaceBrowser)
 from spyder.utils.programs import is_module_installed
