+++ spyder/plugins/variableexplorer/widgets/namespacebrowser.py
@@ -26,7 +26,7 @@ from spyder_kernels.utils.nsview import REMOTE_SETTINGS, get_supported_types
 
 # Local imports
 from spyder.api.translations import get_translation
-from spyder.api.widgets import SpyderWidgetMixin
+from spyder.api.widgets.mixins import SpyderWidgetMixin
 from spyder.widgets.collectionseditor import RemoteCollectionsEditorTableView
 from spyder.plugins.variableexplorer.widgets.importwizard import ImportWizard
 from spyder.utils import encoding
